Increase max throttle speed to 33 gigabits
This commit increases the max throttle speed from 1.6 gigabits to 33 gigabits.

It's no longer possible to use a bit shift operator and we must use some 64 bit unsigned integers.

This is critically useful for ruTorrent to restore throttle functionality, since many connections now exceed 1.6 gigabits.

+++ src/torrent/throttle.h
@@ -54,8 +54,8 @@ public:
   bool                is_throttled();
 
   // 0 == UNLIMITED.
-  uint32_t            max_rate() const { return m_maxRate; }
-  void                set_max_rate(uint32_t v);
+  uint64_t            max_rate() const { return m_maxRate; }
+  void                set_max_rate(uint64_t v);
 
   const Rate*         rate() const;
 
@@ -72,7 +72,7 @@ protected:
   uint32_t            calculate_max_chunk_size() const LIBTORRENT_NO_EXPORT;
   uint32_t            calculate_interval() const LIBTORRENT_NO_EXPORT;
 
-  uint32_t            m_maxRate;
+  uint64_t            m_maxRate;
 
   ThrottleList*       m_throttleList;
 };
